+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 0.0.1
 -----
 
+* [Tools] Nimble
 * [Tools] Danger
 * [Tools] SwiftLint
 * [Added] Array<UInt8>.data -> Data